GW: estimate temporal XG3 means in separate step *
The temporal XG3 means are derived on forehand, on the original scale.
For combinations of type x location x sixyearperiod where all observations are available (6 years),
this is directly calculated. For all other cases (< 6 obs), it is done with
a random effects model.

Note that the actual estimation of variances occurs in the following step,
after log10-transformation.