Design surfw, prep Aquachem: insert file gdb attributes in lazyqry *
This uses the long-formatted location attributes, including a (long)
'type' column, first queried in R from the file geodatabase under
PRJ_Macrofyten.
These location attributes are used in the lazy object after making
it available as a temporary table "##dim_waterhabitat_long" in
Aquachem.

Note that this is a temporary fix. After Aquachem has been further
updated and corrected, this commit should be reverted.
However, the 'type' column in that case (HabTypeVel in FactResultAqua)
will need to be split and put in long format, just as well.