Design gw, MOD 51_aq lg3: drop posterior predictive value *
Although demonstrated by Zuur et al (2017) I think you cannot just compare
observations with the marginals of the fitted values. The latter refer to
the distribution of the (conditional) mean, not of the observations.

This conclusion is confirmed by simulation, hence I think there may be a wrong
implementation in Zuur et al (2017), par 8.6.4 (p. 133).