Design gw, MOD 51_terr lg3: apply RW2-soilclass interaction in all groups *
This change is specifically about group 'GW_05.1_terr_group3_grasslands'.
Before (see 3a6de2a and 8f39463) just an overall RW2 smoother was fitted for this
group, because the computation was not possible in case of fitting
residual heterogeneity & RW2-interaction with the type variable.
However, now that the soilclass variable is used for this, computation
is again OK and the model DIC/WAIC is considerably better.