Design gw, MOD 51_terr lg3: better rw2 prior + diag Hessian *
For subdataset GW_05.1_terr_group3 INLA could not compute
the smarginal() and consequently not the tmarginal()
for the rw2 hyperparameter in the heavy soilclass.
Consequently it has been tried solving this by rectifying
the quite weird PC prior of the rw2 hyperparameters.
However, for this subdataset INLA warned that hyperparameters
could not be trusted when using the regular Hessian matrix.
It was recommended to approach by using the diagonal approximation
instead, of the Hessian.
This has been implemented and estimates are still very similar,
while the post. distibution of the named hyperparameter clearly
improved AND also the WAIC/DIC values improved, sometimes a lot.

So, also intending to backport this fix to 51_aq_lg3 as well!