Design gw, MOD 51_aq lg3: stratify spatial ranef instead of temporal ranef *
Although the fit is worse, this better captures type-specific spatial variance.
It is important for type-specific spatial sample sizes.
More specifically, cluster_id is used to handle this. In order to prevent
correlation between random effects (same as in c19aae3), we use only ONE
spatial random effect, i.e. the most important one: cluster. So loc_id was
dropped from the model.
This means that we ignore spatial dependency between locations inside the
same cluster: this effect now goes into the residuals. They are also type-
specific.
Larger residuals is visible as increased DIC, WAIC and RMSE, so the fit is worse
but the structure better 'fits' the purpose of simulation.

In order to keep the model feasible to fit, we also replace the type-specific temporal
random effect by one common random effect. This does matter for type-specific simulations
of trend, although residual differences between types will still have an effect there.
But state assessment is more important than trend assessment here, so opting
for this route. Long-term trend is still estimated in a type-specific way; so we won't have
drawbacks in detrending.