Explore targetpop, shallowgw: add second layer to shallow_obliggwdep.gpkg *
The second layer is a subset of the first, including only the polygons that
originate from hmt_pol_obliggwdep (i.e. with non-missing polygon_id), and with
the extra column 'areal_frac_obliggwdep_adj' added.
In other words this targets all polygons with obliggwdep types and nothing else,
while documenting the relationship with shallowgroundwater.