Choices_website: move all compiling to render_site.R *
This makes the process of compiling more robust and reproducible.

It entailed a.o.:
- writing render_site.R
- dropping obsolete chunks from detailed docs
- simplifying filepaths in YAML headers and figure references

The script also contains code to generate a PDF version of the site,
and html versions of the detailed documents. This code is not run when
just running the script.

Just running the whole script will create the website with the detailed
documents included as PDF (since they're referred as such from the
downloads section of the website).