Design gw, MOD 51_terr hg3: split lincomb and plot polynomial of binom *
We want to detrend both the binomial and the gamma submodel when doing
simulations, so we need to isolate them both in the model.
When simulating trends, we will probably just adapt the gamma submodel but
it might be cool to somehow connect the random effects of the binomial
and gamma submodels.