Design gw, MOD 51_terr hg3: implement shared random effects *
Considering the correlation between random effects of both submodels, which
violates the model assumption (and would not occur in simulation) but which
is totally sensible, the correlated random effects for hydroyear, loc_code and
cluster_id were now specified as such. See Gómez-Rubio (par. 10.7.5) for a
similar approach. The beta hyperparameter serves as the coefficient to derive
the random intercepts of the gamma submodel from those of the binomial submodel.

It was noted that this largely improves the model evaluation.