Design gw, MOD 51_terr hg3: deliberate choice of stratumvars *
Depending on the typegroup, the best fitting choice of stratumvariable
is made, being either type or soilclass:

- to support heterogeneity in residuals
- for interaction with time in long-term trend, or with space (not applied here)

Contrary to LG3, soilclass is only used for typegroup 3.

Do note that during all previous commits for 51_terr hg3, only the type
stratum variable had been considered because of an uncommitted tweak.
So all comments on results did not consider the soilclass stratum variable.