Design gw, sim*_pops_singlemodel(): modify model-matrix if needed *
It can happen (e.g. in the gamma submodel for HG3 in GW_05.1_terr) that
the model matrix, which is created from the model formula, contains
more columns than corresponding fixed effects in the model summary.
This is because in the model, an effect may have been absorbed in the
artificial intercept.

In that case, it is safe (but also needed) to drop the extra column of
the model matrix before doing the multiplication. Note that the check
of correspondance between model matrix columns and selected fixed
effects is still being done after dropping the extra model-matrix
column.