Design gw, summarise_status_of_samples(): protect against NaN *
quantile() errors when it gets NAs/NaNs as input, so they must be
excluded. The NaNs in this case correspond to relative error margins
for samples where the mean is zero (which comes in the denominator).
This is to be expected for response variables like hg3_lcl, which is
often zero.