Design gw, SIM: aim at both abs & rel error margin for status *
The quality standard for status had to do solely with the relative error margin.
However this measure is rather uninformative in the case of mean values close
to zero, since even with small absolute error margins the relative error margin
tends to be large.
For this reason, also an absolute error margin has been given a quality standard,
at least in GW_05.1_terr.

Consequently, the absolute error margin is now shown systematically in figures and
tables.

Also, the summary object for status (written to disk) is extended and can
now contain the results for > 1 statistic.