Design gw, SIM, summarise_status_of_samples(): further harden P(stat<qual_std) *
Continues c9ebd9a and 5143f5b.

Since integrate() continues to be difficult to generalize over all situations, i.e. to avoid
error with 'the integral is probably divergent' in case of few observations, a simpler and
probably more robust approach is tried instead:

- calculate density = PDF (unchanged)
- calculate CDF from PDF using cumsum
- turn into a CDF function with approxfun

These steps were moved to a dedicated function, that takes a numeric vector and a quantile
from which a left-tail probability is derived.