Design gw, MOD 33 no3: drop month ranef since it has little added value (low variability) *
The variance of the month ranef is consistently very low (very high precision), compared
to other random effects/residuals. Hence it seems to complicate the model too much in the
light of simulations. The interpretation here is that choice of month has very little
effect on the result.

For NO3, this provides a better fit (WAIC/DIC).

Results:

NO3:

DIC: -22881.16
WAIC: -22913.86
Number of equivalent replicates: 5.119

NH4:

DIC: -7583.001
WAIC: -7673.583
Number of equivalent replicates: 5.691

PO4:

DIC: -41293.73
WAIC: -41282.63
Number of equivalent replicates: 5.937

The fit is a bit worse for NH4 & PO4, but there the number of equiv_repl is a bit higher.