Design gw, SIM: use temporal stratif in spatvar calc *
This is only applied for variables for which temporal sampling is envisaged,
i.e. where spatiotemporal status inference will be applicable.
The temporal stratification matches the way in which spatial variance
will be handled in the design-based spatiotemporal statusinference, i.e.
with poststratification over the time variable.

git note: the no3 file of spatvar has been rewritten but didn't change; this means
that I already committed this new state during previous commit of spatvar in 73b4a54...