DCP MBAA: add INBOmd::pdf_report output format *
This commits still neglects the layout of figs & tables in PDF.

The bookdown::html_document2 output format remains in shape however.

The settings for and behaviour of pdf_report depend on a (currently) dev version
of INBOmd (branch 'fixes'); currently that is commit inbo/INBOmd@1a71daa .

The addition of header.tex is to add the underscore TeX package; see
https://stackoverflow.com/questions/2476831. With this, underscores in captions
are automatically supported (no escaping needed).