POC, 010_settings: create distinct revisit designs *
A distinction has been made between field activities (including non-measurement activity)
and spatiotemporal (revisit) design of the variables themselves.

While the latter is important for the statistical underpinning and directs data processing,
the fieldwork is the one we need to prepare data collection plans.

The scheme_design object is dropped and the revisit designs are not directly tied to
schemes anymore.

The link with schemes is scheme_variableset_fieldmethod for variable sets and
scheme_field_activity for field activity; these objects don't include revisit
design information.