POC, 050_sts: fix panel design for SURF field activities *
Before, 12 panels were distinguished per year, while it must the same panel being
repeated 12 times a year. So we had implemented a serially alternating pattern
instead of a pure pattern, within the year.

In the new situation we get many more combinations of locations and time
intervals, because of this panel design.