POC, 050_sts: don't sample same units in domain overlaps *
This adds support for the overlap between domains.

We already had support for identical domains (between modules); here we add
support for a single spatial inclusion relationship between domains.

Chained inclusion relationships are not currently supported.

Partially overlapping domains are not currently supported either: the one
domain should include the other(s).

The sampling approach is as follows, within a scheme x stratum, in case multiple
domains are present that have an inclusion relationship:
- first the larger domain is sampled using the set sample size
- then the smaller domain(s) is/are sampled using the set sample size(s), but for
this only the sampling units are made available that were not yet sampled in
the larger domain. So the smaller domains serve to prolong the GRTS sequence.