POC, 050_sts: fix unit restriction in overlapped domains *
Since modules are not being distinguished here, it is expected that the list of
GRTS addresses for the including domain correspond to a different module
than the included domains.

Consequently, we should not restrict the GRTS addresses of an included
domain to the unsampled GRTS addresses of the including domain, but
instead restrict them by excluding the units that have been sampled in
the including domain.