POC, 010_settings: refactor scheme properties (intermediate state) *
This essentially uses the activity_sequences as a new concept to link activities
to the main datacollection method for a mod x scheme x environmental variable.

This will form the basis in defining panel design, which is still WIP. It is the
aim to define the panel designs in the settings as much as possible; it's not
needed to postpone that to the spatiotemporal sample chapter.

Since we're about to further refactor the underlying data sources, this commit
reflects this intermediate state of those data sources.

Note that, going forward, the fa_fag data source is about to be dropped, since this
connection will be made activity_sequence specific; hence included in the
activity_sequence data source. Also, quite a few activities are about to be
splitted in different (alternative) activities, more activity_sequences will
arise as a consequence, and yet more activities have been distinguished.