Revert "POC, 050_sts: implement limit_n_cycles in panel calendar" *
It has been decided to repeat all field activities according to the cycles.

The reason for 'limit_n_cycles' was for field activities that should normally
happen only once, i.e. at the first ever visit.

However:

- the sample will evolve each cycle, so that new locations replace other ones.
Consequently, those one-time activities will always occur in a new location.
- it is also possible that for some reason the location is still valid (LOCEVAL),
but the activity needs to happen again (the expected long-term validity of criteria
for not repeating the activity does not hold).

Current case were the activities SURFLENTSAMPLPOINT & SURFLOTSAMPLPOINT.

It seems misleading to actually make this a recurring activity in general
(i.e. for each location), but this is solved by (re)defining the activities
in a conditional manner: if the activity has been carried out already, then
it can be dropped. The latter can perhaps be applied at the location level,
i.e. in the spatiotemporal sample, but not at the level of the panel
calendar! But also at the location level this seems not so wise, because it
is still useful to at least check whether the criteria that dismiss the
activity are still valid

This reverts commit f716aff0db05edeabb85cbd301b5ad3408922077.