POC, 050_sts: also implement delays for initial INST & SPATPOSIT panels *
Otherwise, a bunch of panels is dropped in the first cycle because of
month_first_default being larger than 1.

These specific FAGs also cause subsequent activities to be dropped that
depend on it.

Hence it's deemed better to still install and spatially position all
equipment in the first year, regardless of month_first_default.