POC, 050_sts: add various attempts to check actseqs in sps unit cal *
We will immediately reverse this commit, since the checks cannot be
resolved easily. Early on we collapsed activity sequences per FAG for
simplicity, but in order to take FAG order into account, we again
looked at activity sequences. As a consequence, cleaning led to
different results depending on the activity sequence. In the end,
the activity sequence will not be used and disentangling this is
complicated. In order to solve this more easily, we should have kept
FAG x activity sequence separate, which would inflate objects quite a
lot.

Some comments about this will still be added in a later commit.