POC, 050_sts: improve pan memhip for *GAUGE FAGs & drop too-early ones in GW *
SURFINSTGAUGE & SPATPOSITGAUGE FAGs in GW schemes are dropped if they occur earlier
than the GWINST* activities on the same location.

However this is kept to a minimum by applying a 'fastalign' strategy in the
panel membership of these FAGs, meaning that the locations of panel 1 & 13 of GWINST
constitute panel 1 of the GAUGE FAGs, panel 2 & 14 constitutes panel 2 and so on, up
to panel 12. This maximizes matching locations in the same date intervals between
these two different revisit designs of 3 and 6 year periodicity, respectively.