POC, 020_bsf: relax assessed_xxx column *
There's no need to store 'assessed_at_cell_center' since we already filter
the is_cell_center property to be in line with the grts_join method of a type.

Since the cell center aspect can now be directly derived from the type, the
only added value is to mark in the bsf which records are field assessments.
This commit does that.