POC, 030_bsf_proc: also apply phab-sampling to replaced GRTS addresses *
Drawn GRTS addresses in MHQ for cell-center types have not been drawn
using a phab-corrected sampling frame. Keeping them unconditionally,
still introduces the bias which we sought to eliminate in the
phab-correction. Here we apply the phab-correction to those locations
as well, by making use of the phab value of the polygon they belong to.