Add CC-BY license *
Although we keep development (via branches) in the non-public n2khab-mne-design repo,
the main branch will be mirrored in a public n2khab-mne-design1 repo.
To this aim, a license is added already.

We use CC-BY because the repo is mainly composed of the source code of reports and
other authored documents.