POC, compare-obj-script: add section to unravel HQ7140 mystery *
In 8faa6201 a fix was made in determining cross-module popsize of types in MHQ schemes,
when applying a criterion to decide whether a type must be sampled or not. The fix has
caused population size reductions in the application of this criterion, and as a
consequence 7140_meso and 7140_oli were no longer selected to be sampled. Since the target
sample sizes in the pan2124 module have been set at the type level, this now causes a loss
of foreseen sampling units.

A solution is yet to be made!