POC, 020_bsf: fix missingness of certain watersurfaces units *
The points_watsurf object was composed of existing refpoints (including IDs from
watersurfaces_hab) and all other IDs from watersurfaces, but this did not include
all other IDs from watersurfaces_hab. This is because extra polygons from the habitatmap
are not included in watersurfaces.

Going forward, the refpoints should encompass all possible polygons so that this
workaround will not be needed anymore. For now, the new approach meets our requirements,
but the refpoints should become the stable source for the future.