POC, 010: fix (increase) MHQ-pan5 cycle dur & sps size of 3260 *
I just found that the fixes to adjust aquatic MHQ cycle durations from 6 to 3 years, in 8f1024a5 and then
refixed in 81455219, had to exclude 3260 (except in pan2124, where everything is at 3 years). This insight is
based on reading Westra et al (2022): MHQ in Flanders for 3260 is done with a 6 year cycle.