POC, 010+080: don't simplify spsubset to NA *
In this way, we can avoid that actseq values appear in spatial subsets where they
don't belong. Before, NA values of in_aquatic_subset led to more actseqs being
linked to FAGs than would be logical given the type at hand. This leads to no
important changes in results, except for these actseq attributes of FAGs. Also,
it leads to the duplication of 7220 into TERR and AQ subset for GWINST as long as
this is not taken care of explicitly. In the end, this resolves into the exact
same spatial samples and FAG calendar (except when in_aquatic_subset is considered,
which duplicates 7220).