POC: move various eval chunk options to if statements *
This is done to ease interactive usage & debugging of the Rmd files. On
condition that the interactive param settings are OK, Rmd files can be
nog be run as a whole, skipping the same code as what before happened only in
non-interactive compilation. Before, that code needed to be skipped more
carefully when run interactively.