POC, dcpfw-script, org: fix polygon-constrained replacement *
Being too focused on the role of the level 3 cells, I had temporarily
forgotten that we will should implement this _on top_ of polygon-
constrained replacement, and we agreed (2025-03-18) to do this _only_
when the polygon is larger than 64 GRTS cells.

So the polygon-based replacement cells must always be generated, and if
the polygon is too large, they must be restricted by the level 3 cell.

This commit mainly reorders existing code in order to accomplish this.