POC, dcpfw-script, org: add dimension criterion to choose replac cells *
While we agreed to restrict to level 3 replacement cells if the
polygon exceeds 64 cells - the size of a level 3 cell - a polygon
may also be very long and narrow, so that the travel distance to
check on all these cells becomes way to long. Consequently, we
decide to also add a distance criterion, based on the diagonal of
the bounding box of polygon cell centers, which should not be
exceeded.

We currently took the diagonal of a level 3 cell as the allowed
upper limit for this distance criterion. If it is exceeded, then
restriction to the level 3 replacement cells takes place.