POC: change 'main field method' to 'main field activity' *
This is done for consistency with the term 'field activity'.

Main field method was a conversion from 'main data collection method',
limited to the field activities. The term 'main' is solely from the
viewpoint of an (environmental) variable, it has nothing to do with
scheme, nor target variable (it exists for 'any' variable).