POC, 010_settings: provision for split pan des in all MNE schemes *
This is done to maximize synergy among MNE schemes and try to get robust
GRTS partitions that can cope with new modules etc: always allow two
revisit designs, and provision for it now in the GRTS partitions.

Note that we currently don't yet define a different revisit design
between both panel sets of the SURF & SOIL schemes, but this can be
added later. However the 3/4 vs 1/4 proportion of spatial sample
size is implemented everywhere!