POC, 080_sptempsamp: update (fix) condition in GW GAUGE FAG cleanup *
Since panel set 1 of READDIVER FAGs now also has the memship_alignfastest
flag, using this flag to filter the GW GAUGE FAGs is too broad.

The filter has been rewritten to be more specific.

In fact, keeping it was still harmless for READDIVER results, but
it's no longer logical to use it; the READDIVER calendar needs its
own cleanup anyway.