POC, 080_sptempsamp: update READDIVER timing: always some time after INST *
Before, only the READDIVER occasions that had an older date
interval than their corresponding INST FAG were dropped (in a
later step based on FAG ranks), so that those in the same
interval as INST were still kept. All this is now taken care
of in an explicit, early step.