POC, distribute_sample_over_panels(): fix bug in local pivotal sampling *
The implication of the values of the 'auxiliary variables' had been misunderstood
at first. Based on lpm1() documentation and Grafström et al. (2012), a
unit's nearest neighbour(s) is determined based on the values of coordinates
or of other auxiliary variables. Here, we have a single dimension: the panels.
In order to downweight the probability of selecting two adjacent panels, a
unit's 'nearest neighbour(s)' and the distance in between, is determined
based on the variable's values.

The original approach would say that all distances are zero, which likely
leads to simple random sampling.

This commit makes neighbouring panels equidistant with a distance of 1.

The results are indeed much better.