POC, functions.R: experimenting to stabilize pan memship *
The aim is to find a way to make the generic panel membership for consecutive
sample sizes interdependent, i.e. to minimize differences. And store the
distribution for all possible sample sizes per panel_set x stratum x
sp_poststratum.

The approach in this commit sets the inclusion probabilities of all panels
for a given sample size to 1 when drawing the extra panel for sample size + 1.

However, this gives no satisfying result. It appears that the position of the
already drawn panels is neglected in creating the balanced sample, hence new
panels tend to be neighbours of already drawn ones, far too quickly.