POC, functions.R: experimenting to stabilize pan memship: GRTS *
According to foregoing results, the implementation
of local pivotal sampling in the BalancedSampling
package seems limited in terms of capabilities to
sample adaptively using LPM1, i.e. keeping already
sampled panels when increasing sample size.

Consequently, we prefer to generate GRTS addresses in
one dimension in order to set the order in which
panels can be sampled. This means abandoning local
pivotal sampling.